Extract log event processing from TelemetryAppenderImpl.
This is useful for testing
