+++ lib/java/logging-service-telemetry/src/main/java/org/enso/logging/service/telemetry/LogJobsQueue.java
@@ -1,18 +1,17 @@
 package org.enso.logging.service.telemetry;
 
-import ch.qos.logback.classic.spi.ILoggingEvent;
 import java.util.ArrayList;
 import java.util.Deque;
 import java.util.LinkedList;
 import java.util.List;
 
 final class LogJobsQueue {
-  private final Deque<ILoggingEvent> queue = new LinkedList<>();
+  private final Deque<LogMessage> queue = new LinkedList<>();
 
   /** Enqueues a log message to be sent and returns the number of messages in the queue. */
-  synchronized int enqueue(ILoggingEvent job) {
+  synchronized int enqueue(LogMessage message) {
     int previousSize = queue.size();
-    queue.addLast(job);
+    queue.addLast(message);
     int newSize = queue.size();
     assert newSize == previousSize + 1
         : "Appending to queue is synchronized, so the size is always incremented exactly by 1.";
@@ -20,14 +19,14 @@ final class LogJobsQueue {
   }
 
   /** Removes and returns up to {@code limit} enqueued jobs. */
-  synchronized List<ILoggingEvent> popEnqueuedJobs(int limit) {
+  synchronized List<LogMessage> popEnqueuedJobs(int limit) {
     assert limit > 0;
     if (queue.isEmpty()) {
       return List.of();
     }
 
     int n = Math.min(limit, queue.size());
-    List<ILoggingEvent> result = new ArrayList<>(n);
+    List<LogMessage> result = new ArrayList<>(n);
     for (int i = 0; i < n; i++) {
       result.add(queue.removeFirst());
     }
